src/requirements.txt
pymongo>=4.0
python-dotenv>=1.0
google-cloud-bigquery>=3.0
